Move MockDevice and get_restore_config into phone.utils
