+++ corehq/apps/ota/tests/test_app_aware_restore.py
@@ -138,7 +138,7 @@ class AppAwareSyncTests(TestCase):
         self.assertEqual(len(fixtures), 1)
 
     def test_user_restore(self):
-        from casexml.apps.phone.tests.utils import MockDevice
+        from casexml.apps.phone.utils import MockDevice
         from casexml.apps.case.xml import V3
         from corehq.apps.userreports.reports.data_source import ConfigurableReportDataSource
