+++ corehq/ex-submodules/casexml/apps/case/tests/test_sync_purge.py
@@ -5,7 +5,8 @@ from couchdbkit.exceptions import ResourceNotFound
 from django.test import TestCase
 
 from casexml.apps.case.xml import V1
-from casexml.apps.phone.tests.utils import create_restore_user, MockDevice
+from casexml.apps.phone.tests import create_restore_user
+from casexml.apps.phone.utils import MockDevice
 from corehq.apps.domain.models import Domain
 from corehq.apps.receiverwrapper.util import submit_form_locally
 from corehq.form_processor.tests.utils import FormProcessorTestUtils
