+++ corehq/ex-submodules/casexml/apps/phone/tests/test_new_sync.py
@@ -18,9 +18,9 @@ from casexml.apps.phone.models import (
 from casexml.apps.phone.restore import RestoreConfig
 from casexml.apps.phone.tests.utils import (
     create_restore_user,
-    MockDevice,
     deprecated_synclog_from_restore_payload,
 )
+from casexml.apps.phone.utils import MockDevice
 from corehq.apps.domain.models import Domain
 from corehq.form_processor.tests.utils import use_sql_backend
 from corehq.toggles import LEGACY_SYNC_SUPPORT
