+++ corehq/ex-submodules/casexml/apps/phone/tests/test_ota_restore.py
@@ -3,10 +3,8 @@ from datetime import datetime
 from django.test import TestCase
 import os
 from django.test.utils import override_settings
-from casexml.apps.phone.tests.utils import (
-    deprecated_generate_restore_payload,
-    get_restore_config,
-)
+from casexml.apps.phone.tests.utils import deprecated_generate_restore_payload
+from casexml.apps.phone.utils import get_restore_config
 from corehq.apps.receiverwrapper.util import submit_form_locally
 from casexml.apps.case.tests.util import check_xml_line_by_line, delete_all_cases, delete_all_sync_logs, \
     delete_all_xforms
