+++ corehq/ex-submodules/casexml/apps/phone/tests/test_ota_restore_v3.py
@@ -10,7 +10,8 @@ from casexml.apps.case.tests.util import (
 )
 from casexml.apps.case.mock import CaseBlock
 from casexml.apps.phone.restore import RestoreContent
-from casexml.apps.phone.tests.utils import create_restore_user, MockDevice
+from casexml.apps.phone.tests.utils import create_restore_user
+from casexml.apps.phone.utils import MockDevice
 
 
 @override_settings(CASEXML_FORCE_DOMAIN_CHECK=False)
