+++ corehq/ex-submodules/casexml/apps/phone/tests/test_sync_logs.py
@@ -5,7 +5,8 @@ from casexml.apps.case.xml import V1
 from casexml.apps.phone.models import SyncLog, CaseState
 from casexml.apps.case.sharedmodels import CommCareCaseIndex
 from casexml.apps.phone.restore import RestoreParams, RestoreConfig
-from casexml.apps.phone.tests.utils import create_restore_user, MockDevice
+from casexml.apps.phone.tests.utils import create_restore_user
+from casexml.apps.phone.utils import MockDevice
 from corehq.apps.app_manager.models import Application
 from corehq.apps.domain.models import Domain
